SCP-5085 Simple contract list view working
This is a first attempt at this, displaying ALL Marlowe contracts that
come back from the REST call with no paging or automatic refreshing.

The contract links drill down into contract details and a link is
available from the contract tab pages to navigate back to this list.

Some work was done to refactor some types used for parsing the JSON.
More work of this type is coming soon.

Also found an extra import hiding in Explorer.Web.ContractView