Factored out the Link concept into a reusable type
This value is used several places but we always see a single useful
value in it. This code provides a Link type we can use everywhere
provided we're explicit about how it's parsed from the JSON in each
case (the JSON paths vary i.e. links.contract vs links.transactions)